Add lifecycle state to base classes
