+++ src/models/base/index.ts
@@ -1,0 +1,3 @@
+export * from './base-component';
+export * from './base-relationship';
+export * from './enums';
